+++ ng-workspace/projects/OpenWaterFoundation/common/ui/dashboard/dashboard.component.ts
@@ -93,7 +93,9 @@ export class DashboardComponent implements OnDestroy {
    * Called once, before the instance is destroyed.
    */
   ngOnDestroy(): void {
-    this.dashboardConfigPathSub$.unsubscribe();
+    if (this.dashboardConfigPathSub$) {
+      this.dashboardConfigPathSub$.unsubscribe();
+    }
   }
 
   /**
@@ -106,7 +108,7 @@ export class DashboardComponent implements OnDestroy {
     // If no style object is provided, return the default object.
     if (!style) {
       return {
-        backgroundColor: 'gray'
+        backgroundColor: 'lightgrey'
       }
     }
 
@@ -129,7 +131,7 @@ export class DashboardComponent implements OnDestroy {
     // The property does not exist, so return a default value.
     else {
       switch (style) {
-        case IM.Style.backgroundColor: return 'gray';
+        case IM.Style.backgroundColor: return 'lightgrey';
         case IM.Style.color: return 'black';
       }
     }