+++ ng-workspace/projects/OpenWaterFoundation/common/ui/dashboard/widget/chart/chart-widget.component.ts
@@ -3,11 +3,6 @@ import { Component,
           OnDestroy, 
           OnInit}             from '@angular/core';
 
-import { BehaviorSubject,
-          Observable }        from 'rxjs';
-
-
-import { OwfCommonService }   from '@OpenWaterFoundation/common/services';
 import * as IM                from '@OpenWaterFoundation/common/services';
 
 
@@ -18,33 +13,15 @@ import * as IM                from '@OpenWaterFoundation/common/services';
 })
 export class ChartWidgetComponent implements OnInit, OnDestroy {
 
-  /** Set to true if any errors occur in the Chart Widget. */
-  private chartError: BehaviorSubject<boolean> = new BehaviorSubject(false);
   /** The attribute provided to this component when created, e.g.
    *   <chart-widget [chartWidget]="widget"></chart-widget> */
   @Input() chartWidget: IM.ChartWidget;
-  /** String array representing the type of error that occurred while building this
-   * widget. Used by the error widget. */
-  errorTypes: string[] = [];
 
 
   /**
   * @constructor for the DialogTSGraph Component.
-  * @param commonService A reference to the top level service OwfCommonService.
   */
-  constructor(private commonService: OwfCommonService) { }
-
-
-  /**
-   * Observable used in the Chart Widget that's used with an async pipe in the template
-   * file to show the widget or error content.
-   */
-  get isChartError$(): Observable<boolean> { return this.chartError.asObservable(); }
-
-  /**
-   * Toggles the chartError BehaviorSubject between true and false.
-   */
-  set setChartError(error: boolean) { this.chartError.next(error); }
+  constructor() { }
 
 
   /**